refactor: moved class declarations to models folder; updated swagger spec
