+++ api/models/feedback_model.py
@@ -1,0 +1,5 @@
+# Schema for Feedback object
+class FeedbackModel:
+    def __init__(self,description, url):
+        self.description = description
+        self.url = url
