+++ api/models/links_model.py
@@ -1,0 +1,5 @@
+# Schema for links object
+class LinksModel():
+    def __init__(self, id):
+        self.self = f"/bids/{id}"
+        self.questions = f"/bids/{id}/questions"
