+++ app.py
@@ -21,4 +21,4 @@ app.register_blueprint(bid, url_prefix='/api')
 
 
 if __name__ == '__main__':
-    app.run(debug=True, port=3000)
+    app.run(debug=True, port=8080)
