--- db.txt
@@ -1,1 +1,0 @@
-{"status": "in_progress", "alias": "ONS", "bid_date": "2023-06-21", "client": "Office for National Statistics", "tender": "Business Intelligence and Data Warehousing", "last_updated": "2023-06-30T14:43:27.022464", "feedback": {"description": "Feedback from client in detail", "url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback"}, "id": "6cb57630-5fd6-404a-a29f-a1aaf821fd61", "links": {"self": "/bids/6cb57630-5fd6-404a-a29f-a1aaf821fd61", "questions": "/bids/6cb57630-5fd6-404a-a29f-a1aaf821fd61/questions"}, "was_successful": false, "success": [{"phase": 1, "has_score": true, "score": 28, "out_of": 36}, {"phase": 2, "has_score": false}], "failed": {"phase": 3, "has_score": true, "score": 20, "out_of": 36}, "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder"}{"id": "2bf95109-ffe5-4b69-abbd-6f356f1f7103", "links": {"self": "/bids/2bf95109-ffe5-4b69-abbd-6f356f1f7103", "questions": "/bids/2bf95109-ffe5-4b69-abbd-6f356f1f7103/questions"}, "bid_date": "2023-06-21", "tender": "Business Intelligence and Data Warehousing", "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder", "was_successful": false, "failed": {"score": 20, "out_of": 36, "has_score": true, "phase": 3}, "client": "Office for National Statistics", "feedback": {"url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback", "description": "Feedback from client in detail"}, "status": "in_progress", "last_updated": "2023-06-30T14:57:52.772927", "alias": "ONS", "success": [{"score": 28, "out_of": 36, "has_score": true, "phase": 1}, {"has_score": false, "phase": 2}]}{"client": "Office for National Statistics", "status": "in_progress", "id": "274083b5-d660-467f-a58a-243941eac574", "alias": "ONS", "links": {"self": "/bids/274083b5-d660-467f-a58a-243941eac574", "questions": "/bids/274083b5-d660-467f-a58a-243941eac574/questions"}, "failed": {"has_score": true, "out_of": 36, "phase": 3, "score": 20}, "bid_date": "2023-06-21", "success": [{"has_score": true, "out_of": 36, "phase": 1, "score": 28}, {"has_score": false, "phase": 2}], "feedback": {"url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback", "description": "Feedback from client in detail"}, "bid_folder_url": "test", "tender": "Business Intelligence and Data Warehousing", "last_updated": "2023-06-30T15:00:40.137631", "was_successful": false}{"alias": "ONS", "id": "ba083965-3737-4fad-8a8d-6a40cbcab5bb", "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder", "client": "Office for National Statistics", "bid_date": "2023-06-21", "last_updated": "2023-06-30T15:13:13.123372", "was_successful": false, "feedback": {"url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback", "description": "Feedback from client in detail"}, "links": {"self": "/bids/ba083965-3737-4fad-8a8d-6a40cbcab5bb", "questions": "/bids/ba083965-3737-4fad-8a8d-6a40cbcab5bb/questions"}, "tender": "Business Intelligence and Data Warehousing", "failed": {"score": 20, "has_score": true, "phase": 3, "out_of": 36}, "status": "in_progress", "success": [{"score": 28, "has_score": true, "phase": 1, "out_of": 36}, {"has_score": false, "phase": 2}]}{"alias": "ONS", "id": "00072bff-cb8e-45ef-ab09-c7860b719bf2", "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder", "client": "Office for National Statistics", "bid_date": "2023-06-21", "last_updated": "2023-06-30T15:15:34.931648", "was_successful": false, "feedback": {"url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback", "description": "Feedback from client in detail"}, "links": {"self": "/bids/00072bff-cb8e-45ef-ab09-c7860b719bf2", "questions": "/bids/00072bff-cb8e-45ef-ab09-c7860b719bf2/questions"}, "tender": "Business Intelligence and Data Warehousing", "failed": {"score": 20, "has_score": true, "phase": 3, "out_of": 36}, "status": "in_progress", "success": [{"score": 28, "has_score": true, "phase": 1, "out_of": 36}, {"has_score": false, "phase": 2}]}
