+++ request_examples/invalid_url.http
@@ -1,4 +1,4 @@
-POST http://localhost:3000/api/bids HTTP/1.1
+POST http://localhost:8080/api/bids HTTP/1.1
 Content-Type: application/json
 
 {
