+++ static/swagger_config.yml
@@ -22,7 +22,7 @@ info:
 # --------------------------------------------
 # Server
 servers:
-  - url: 'http://localhost:3000/api/'
+  - url: 'http://localhost:8080/api/'
     description: Local server
 # --------------------------------------------
 # Tags
@@ -195,12 +195,12 @@ components:
     QuestionsLink:
       description: A link to a collection of questions for a bid
       type: string
-      example: 'https://localhost:3000/api//bids/96d69775-29af-46b1-aaf4-bfbdb1543412/questions'
+      example: 'https://localhost:8080/api//bids/96d69775-29af-46b1-aaf4-bfbdb1543412/questions'
 # --------------------------------------------
     SelfLink:
       description: A link to the current resource
       type: string
-      example: 'https://localhost:3000/api/bids/471fea1f-705c-4851-9a5b-df7bc2651428'
+      example: 'https://localhost:8080/api/bids/471fea1f-705c-4851-9a5b-df7bc2651428'
 # --------------------------------------------
     BidRequestBody:
       type: object
@@ -258,6 +258,25 @@ components:
                 client: 'Office for National Statistics'
                 alias: 'ONS'
                 bid_date: '21-06-2023'
+                success: [
+                  {
+                    "phase": 1,
+                    "has_score": true,
+                    "score": 28,
+                    "out_of": 36
+                  },
+                  {
+                    "phase": 2,
+                    "has_score": false
+                  }
+                ]
+                failed: {
+                  "phase": 3,
+                  "has_score": true,
+                  "score": 20,
+                  "out_of": 36
+                }
+                was_successful: false
             400 Bad Request:
               summary: 400 Bad Request
               value:
@@ -274,15 +293,19 @@ components:
     BadRequest:
       description: Bad Request Error
       content:
-        text/plain:
+        application/json:
           schema:
-            type: string
-            example: Missing mandatory field
+            type: object
+            example: {
+              "Error": "{'<field>': ['<message>']}"
+            }
     InternalServerError:
       description: Internal Server Error
       content:
-        text/plain:
+        application/json:
           schema:
-            type: string
-            example: Could not connect to a database
+            type: object
+            example: {
+              "Error": "{'<field>': ['<message>']}"
+            }
 # --------------------------------------------
